Analysis of formulas in the queue
The file analysis have the code to parse the txt file with the evolution of the queue and calculate variables and formulas. The file quantidades describes the quantities required in the work. The file fila is a example generated by main.py. The test file wallpaper inserted by me as test of gitwas removed.